Refactored and reorganized code to make it modular
